requirements.txt
Django==3.2
django-cors-headers==3.11.0
django-filter==21.1
django-rest-swagger==2.2.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.1.0
Jinja2==3.1.1
Markdown==3.3.6
openapi-codec==1.3.2
Pillow==9.0.1
psycopg2-binary==2.9.3
PyJWT==2.3.0
setuptools==58.1.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9